requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
assemblyai==0.35.1
asyncpg==0.30.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
databases==0.9.0
elevenlabs==1.50.2
et_xmlfile==2.0.0
fastapi==0.115.6
frozenlist==1.5.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
loguru==0.7.3
multidict==6.1.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
twilio==9.4.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
yarl==1.18.3
